+++ perk_creator.py
@@ -532,7 +532,7 @@ def handle_adding_perks_to_vanilla_nodes():
         ("powerful1",1),
         ("fury1",1),
         ("mark1",3),
-        ("fortify1",1),
+        ("fortify1",2),
         ("sharp1",1),
         ("energy2",1),
         ("insulate1",2),