+++ PerksCSharp/PerkProject/Plugin.cs
@@ -165,7 +165,7 @@ namespace PerkManager{
             
             // Custom Text for Fortify
             medsTexts[perkStem + "fortify1d"] = "At the end of this hero's turn, gain 1 Reinforce for every 2 Fortify charges.";
-            medsTexts[perkStem + "fortify1e"] = "Fortify on all heroes is capped at 5, but reduces damage taken by 1 per charge.";
+            medsTexts[perkStem + "fortify1e"] = "Fortify on all heroes can stack up to 5, but reduces damage taken by 1 per charge.";
 
             // Custom Text for Sharp
             medsTexts[perkStem + "sharp1e"] = "If Sharp on a hero would increase a damage type, it increases it by 1.5 damage per charge. Sharp on heroes only stacks to 25.";
@@ -178,7 +178,8 @@ namespace PerkManager{
             medsTexts[perkStem + "crack2e"] = "Crack on monsters reduces Lightning resist by 0.3% per charge.";
             medsTexts[perkStem + "crack2f"] = "Crack on monsters increases Fire damage received by 1 per charge in addition to Blunt damage.[Not compatible with the Mind version.]";
             medsTexts[perkStem + "crack2g"] = "Crack on monsters increases Mind damage received by 1 per charge in addition to Blunt damage. [Overrides the Burn version.]";
-            medsTexts[perkStem + "crack2h"] = "Crack on monsters reduces Lightning resist by 0.3% per charge.";
+            medsTexts[perkStem + "crack2h"] = "Crack on monsters reduces Slashing resist by 0.3% per charge.";
+            medsTexts[perkStem + "crack2i"] = "Crack on monsters increases max Vulnerable charges by 1 per 25 charges of Crack.";
 
             // Custom Text for Burn
             medsTexts[perkStem + "burn2e"] = "Burn increases the damage dealt by Dark explosions by 0.5% per charge.";