+++ Too Many Perks/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Updated: Energy2c, Bleed2b, Bleed2c, Sharp1d,
 
-Added: Mark1f, 
+Added: Mark1f, Crack2h
 
 # 1.2.3