+++ Too Many Perks/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 Updated: Energy2c, Bleed2b, Bleed2c, Sharp1d, Burn2d, Spark2c, Reinforce1b, Insulate1b, Courage1b
 
-Added: Mark1f, Crack2h
+Added: Mark1f, Crack2h  
+
+These have been minimally tested, so please report bugs with them.
 
 # 1.2.3